Added sample configs where config files could not be included
Added a sample config for aws.php and config.php as those files contain
sensitive information and need to be set per install.
The database.php config file is now included as this project does not
depend on a database.
The index.php file is included, it defaults to a production environment
and has been enhanced to insure Composer has been run and set up
correctly.